add results route
